+++ scm-core/src/br/cefetmg/inf/model/service/GenericManter.java
@@ -8,7 +8,7 @@ public interface GenericManter<Entidade> {
     
     public void cadastrar(Entidade e) throws PersistenciaException, NegocioException;
     public void alterar(Entidade e) throws PersistenciaException, NegocioException;
-    public Entidade buscarPorId(Long id) throws PersistenciaException, NegocioException;
+    public Entidade buscarPorId(Integer id) throws PersistenciaException, NegocioException;
     public void excluir(Entidade e) throws PersistenciaException, NegocioException;
     public List<Entidade> listarTodos() throws PersistenciaException, NegocioException;
