+++ scm-core/src/br/cefetmg/inf/util/db/persistence/GenericDAO.java
@@ -5,10 +5,10 @@ import java.util.List;
 
 public interface GenericDAO<Entidade> {
 
-    Long inserir(Entidade obj) throws PersistenciaException;
+    Integer inserir(Entidade obj) throws PersistenciaException;
     void atualizar(Entidade obj) throws PersistenciaException;
-    void excluir(Long id) throws PersistenciaException;
-    Entidade consultarPorId(Long id) throws PersistenciaException;
+    void excluir(Integer id) throws PersistenciaException;
+    Entidade consultarPorId(Integer id) throws PersistenciaException;
     List<Entidade> listarTodos() throws PersistenciaException;
 
 }
